Move duplicated methods for packages to parent controller
In order to make the codebase a little more DRY, methods
used across the packages sub-controllers can be moved to a
parent controller.
